requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
Flask-Login==0.6.3
psutil==5.9.5
python-engineio==4.8.0
python-socketio==5.10.0
python-telegram-bot==20.8
flask-sqlalchemy==3.1.1
werkzeug==3.0.1
python-dotenv==1.0.1
